requirements.txt
Pillow==9.0.1
xtermcolor~=1.3
nuitka
PyYAML~=6.0